requirements.txt
baidu-aip
pyaudio
wave
